Account delete
account delete complete

+++ users/views.py
@@ -5,7 +5,7 @@ from .models import CustomUser
 from django.db import connection
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth import login,authenticate
-from django.contrib.auth.decorators import user_passes_test
+from django.contrib.auth.decorators import login_required
 from django.urls import reverse_lazy
 class SignUpView():
     def signup_view(request):
@@ -27,15 +27,21 @@ class UserDetailView(LoginRequiredMixin,DetailView):
     def get_object(self):
         return self.request.user
 
-@user_passes_test(lambda u: u.is_staff, login_url=reverse_lazy('login'))
-class DeleteView(LoginRequiredMixin):
-    def delete_user(request):
-        context = {}
-        print(request.user)
+
+@login_required(login_url='home.html')
+def delete_account(request):
+    context = {}
+    print(request.user)
+    password = request.POST["password"]
+    user=CustomUser.objects.get(username=request.user)
+    if request.user.check_password(password):
+        print("yes")
         try:
             u = CustomUser.objects.get(username=request.user)
             u.delete()
         except Exception as e:
             print(e)
-            return render(request,'delete_account.html',{'Exception':e})
+            return render(request,'accounts/delete_account.html',{'Exception':e})
         return render(request,'accounts/delete_account_success.html')
+    e="Password is not correct."
+    return render(request,'accounts/delete_account.html',{'Exception':e})
